% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_labelling.R
\name{batch_labelling}
\alias{batch_labelling}
\title{Batch process a set of target analytes}
\usage{
batch_labelling(targets, groups, plot_patterns = T, plot_residuals = F,
  plot_results = F, save_results = F)
}
\arguments{
\item{targets}{A data frame containing information on the target analytes. Columns 1 to 8 are: "name_compound" (the name of the target analytes), "compound", "charge", "labelling", "RT", "RT_shift", "chrom_width", "mass_shift" (the same parameters for \code{\link{main_labelling}} function). Columns from 9 onwards contain the initial estimates for the label abundances (one estimate for each sample). If not known, a single column of NA values can be entered}

\item{groups}{A factor containing the name of the group of each sample analysed; 
The function will calculate summary statistics for the samples belonging to the same group}

\item{plot_patterns, plot_residuals, plot_results}{Whether or not to plot the patterns, the residuals and a summary of the results. If so, pdf files are created in the working directory}

\item{save_results}{Whether to save the results of the estimates. If so, *.csv files are generated}
}
\value{
batch_grouped_estimates A list as long as the number of target analytes, containing the summary of the fitted results (group estimates)
}
\description{
This function batch processes LC- or GC-MS data, analyzing the isotopic patterns of a set of target analytes specified by the user. As with other functions of the package, it also requires chromatogrpahic information.
}
\examples{
# Get the sample dataset
data("xcms_obj")

# Convert the MS data set
peak_table <- table_xcms(xcms_obj)

# Get the example data frame containing target abalytes
data("targets")

 # Batch process the data
 batch_grouped_estimates <- batch_labelling(targets=targets, 
 groups=factor(c(rep("C12",4), rep("C13",4))),
 plot_patterns=FALSE, plot_residuals=FALSE, plot_results=FALSE, save_results=FALSE)
   
}
\author{
Ruggero Ferrazza
}
\seealso{
\link{main_labelling}, \link{group_labelling}, \link{save_labelling}, \link{plot.labelling}
}

