\name{labelling-class}
\alias{labelling}

\title{
Class labelling
}

\description{
A class \code{labelling} object is a list containing the results of the isotopic pattern analysis. Such object is the output of the functions \code{\link{find_abundance}} and \code{\link{main_labelling}}.
}

\details{
The methods available for class "labelling" objects are:
\describe{
	\item{summary}{to produce a summary of the results contained in the \code{labelling} object;}
	\item{plot}{to create graphical outputs.}
}

}

\seealso{
\code{\link{find_abundance}}, \code{\link{main_labelling}}, \code{\link{summary}}, \code{\link{plot}}
}