% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.labelling.R
\name{summary.labelling}
\alias{summary.labelling}
\title{Summary of the labelling fitting}
\usage{
\method{summary}{labelling}(object, ...)
}
\arguments{
\item{object}{Object of class \code{labelling},
output of either \code{\link{main_labelling}} or \code{\link{find_abundance}} functions}

\item{...}{Additional parameters}
}
\value{
\item{results}{Matrix containing a summary of the fitted results. It has two rows, 
the first containing the estimated percentage isotopic abundances of the labelling isotope X (^2H or ^13C), 
and the second one containing the standard errors from the fitting procedure}
}
\description{
Function that produces a summary of the results from an object of class \code{labelling}.
}
\author{
Ruggero Ferrazza
}
\keyword{manip}

