% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_xcms.R
\name{table_xcms}
\alias{table_xcms}
\title{Process \code{xcmsSet}}
\usage{
table_xcms(xcms_obj)
}
\arguments{
\item{xcms_obj}{an xcmsSet object}
}
\value{
\item{peak_table}{data frame extracted from the \code{xcmsSet} object. 
The first two columns representing mass and retention 
time of the related peaks}
}
\description{
Function that properly convert an xcmsSet object, 
from package xcms, into a table of peaks
}
\note{
The output data frame, required by other functions of the 
\code{\link{IsotopicLabelling}} R package, 
can be obtained in a number of other independent ways, 
such as through proprietary software of the vendor of the MS instrument.
}
\examples{
data(xcms_obj)
peak_table <- table_xcms(xcms_obj) 

}
\author{
Ruggero Ferrazza
}
\keyword{manip}

