% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcms_obj.R
\docType{data}
\name{xcms_obj}
\alias{xcms_obj}
\title{Example data set}
\format{an \code{xcmsSet} object from the package \code{xcms}}
\source{
Department of Biochemistry, University of Cambridge - UK
}
\usage{
xcms_obj
}
\description{
An "xcmsSet" object containing the results of the LC-MS analysis of 8 lipid extract 
obtained in a ^13C isotopic labelling experiment. 
The first 4 samples were obtained from cell cultures grown under normal conditions (natural ^13C abundance), 
whereas in the remaining 4 the cells were grown using uniformly-labelled 99\% ^13C glucose.
}
\examples{
data("xcms_obj")
}
\author{
Dr. Julian L Griffin and Dr. Nyasha Munjoma
}
\keyword{datasets}

