\name{labeling-class}
\alias{labeling}

\title{
Class labeling
}

\description{
A class \code{labeling} object is a list containing the results of the isotopic pattern analysis. Such object is the output of the functions \code{\link{find_abundance}} and \code{\link{main_labeling}}.
}

\details{
The methods available for class "labeling" objects are:
\describe{
	\item{summary}{to produce a summary of the results contained in the \code{labeling} object;}
	\item{plot}{to create graphical outputs.}
}

}

\seealso{
\code{\link{find_abundance}}, \code{\link{main_labeling}}, \code{\link{summary}}, \code{\link{plot}}
}