% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.labelling.R
\name{plot.labelling}
\alias{plot.labelling}
\title{Plot method for \code{labelling} objects}
\usage{
\method{plot}{labelling}(x, type = "patterns", saveplots = F, ...)
}
\arguments{
\item{x}{Object of class \code{labelling}}

\item{type}{The type of output produced. Available options are "patterns", "residuals", "summary"}

\item{saveplots}{Should the plots be saved as a PDF file?}

\item{...}{Other parameters}
}
\value{
One or more plots
}
\description{
Produces different types of summary plots for a \code{labelling} object.
}
\details{
The default (type 'patterns') plot shows, for each sample in the class \code{labelling} object, 
the normalized experimental pattern superimposed to its fitted theoretical pattern.
By setting type to 'residuals', the function plots the residuals (the differences between experimental and best fitted theoretical patterns).
Type 'summary' produces a summary plot showing the estimated percentage abundances with related standard errors.
If \code{saveplot} is TRUE, the plots are saved to a PDF file in the working directory.
}
\author{
Ruggero Ferrazza
}
\keyword{hplot}

