% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_labelling.R
\name{save_labelling}
\alias{save_labelling}
\title{Export to csv}
\usage{
save_labelling(fitted_abundances, path = getwd())
}
\arguments{
\item{fitted_abundances}{Object of class \code{labelling}}

\item{path}{The directory where to save the csv file. 
If not specified, the results are saved in the working directory}
}
\value{
The "COMPOUND_Estimated_Abundances.csv" file, containing the results of the analysis. 
For each sample (one for each row) there are four columns:
\enumerate{
\item The estimated percentage abundance of the labelling isotope (either ^2H or ^13C);
\item The related standard error coming from the \code{nls} fitting procedure;
\item The percentage deviation between theoretical and experimental isotopic patterns;
\item The outcome message from the fitting procedure, to undersand whether there have been any convergence problems.
}
}
\description{
Function that saves the obtained results to a csv file.
}
\examples{

}
\author{
Ruggero Ferrazza
}
\seealso{
\link{main_labelling}
}
\keyword{IO}

