\name{xcms_obj}
\alias{xcms_obj}
\docType{data}

\title{
Example data set, a class "xcmsSet" object
}

\description{
This data set is an example of LC-MS experimental data obtained following isotopic labeling experiments. The data was kindly provided by Dr. Jules Griffin and Dr. Nyasha Munjoma (Department of Biochemistry, University of Cambridge - UK).
  
It contains LC-MS data relative to \eqn{^{13}}C labeling experiments on 8 samples: the first 4 represent lipid extracts from cell cultures grown under normal conditions (natural \eqn{^{13}}C abundance), whereas in the last 4 samples the lipids were extracted from cells grown using uniformly-labeled \eqn{^{13}}C glucose (99\% \eqn{^{13}}C labeling).

}

\usage{data("xcms_obj")}